improve lineage load time
