+++ web/client/src/library/components/plan/PlanChangePreview.tsx
@@ -183,7 +183,6 @@ function PlanChangePreviewDirect({
                         <div className="h-[16rem] bg-theme-lighter rounded-xl p-2">
                           <LineageFlowProvider withColumns={false}>
                             <ModelLineage
-                              key={model.id}
                               model={model}
                               highlightedNodes={{
                                 'border-4 border-secondary-500': [model.name],
